# API/management/api.yaml
openapi: 3.0.0
info:
  title: ManagementEnd API
  version: 1.0.0
  description: 系统ManagementEnd组件的API
servers:
  - url: https://api.managementend.example.com/v1
tags:
  - name: 同步
    description: 用于同步各种数据的端点
  - name: 配置
    description: 用于管理用户配置的端点
  - name: 文件操作
    description: 用于文件和目录管理的端点
  - name: 任务管理
    description: 用于管理和监控任务的端点
  - name: 系统监控
    description: 用于监控系统资源的端点
paths:
  /file/strategy:
    get:
      tags:
        - 同步
      summary: 请求同步策略注册表
      parameters:
      - name: project_name
        in: query
        required: true
        schema:
            type: string
      responses:
        '200':
          description: 同步成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyRegistryResponse'
        '501':
          description: ErrorResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /file/directory:
    get:
      tags:
        - 同步
      summary: 请求同步项目文件目录
      parameters:
      - name: project_name
        in: query
        required: true
        schema:
            type: string
      responses:
        '200':
          description: 同步成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectoryResponse'
        '501':
          description: ErrorResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /file/project-list:
    get:
      tags:
        - 同步
      summary: 请求项目列表
      responses:
        '200':
          description: 同步成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '501':
          description: ErrorResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /file/config:
    parameters:
      - name: project_name
        in: query
        required: true
        schema:
          type: string
          description: 项目名
    get:
      tags:
        - 配置
      summary: 请求同步配置文件
      responses:
        '200':
          description: 同步成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationResponse'
        '501':
          description: ErrorResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - 配置
      summary: 发送修改后的配置表单
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/ConfigurationRequest'
      responses:
        '200':
          description: SuccessResponse

        '500':
          description: ErrorResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /file:
    parameters:
      - name: file_path
        in: query
        required: true
        schema:
          type: string
        description: 文件路径
    post:
      tags:
        - 文件操作
      summary: 创建文件或文件夹
      responses:
        '200':
          description: SuccessResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '500':
          description: ErrorResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - 文件操作
      summary: 修改文件内容
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileModificationRequest'
      responses:
        '200':
          description: SuccessResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '501':
          description: ErrorResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: ErrorResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - 文件操作
      summary: 查看文件内容
      responses:
        '200':
          description: 成功获取文件内容
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileContentResponse'
        '404':
          description: NotFoundResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: ErrorResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - 文件操作
      summary: 重命名用户文件
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
              description: 新文件名
      responses:
        '200':
          description: SuccessResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '500':
          description: ErrorResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - 文件操作
      summary: 删除用户文件
      responses:
        '200':
          description: SuccessResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '500':
          description: ErrorResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /mission:
    post:
      tags:
        - 任务管理
      summary: 通过 WebSocket 连接发送任务信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionRequest'
      responses:
        '200':
          description: 成功发送任务信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissionResponse'
        '500':
          description: ErrorResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - 任务管理
      summary: 停止运行中的任务或删除等待中的任务
      parameters:
        - in: query
          name: project_name
          schema:
            type: string
          required: true
      responses:
        '200':
          description: SuccessResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '500':
          description: ErrorResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /monitor:
    get:
      tags:
        - 系统监控
      summary: 获取系统状态信息（允许空值）
      responses:
        '200':
          description: 成功获取系统状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStatusResponse'
        '500':
          description: ErrorResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    StrategyRegistryResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
          properties:
            strategy_registry:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/StrategyInfo'
    StrategyInfo:
      type: object
      properties:
        argus:
          type: array
          items:
            type: object
            properties:
              argu_name:
                type: string
              argu_annotation:
                type: string
              argu_default:
                type: string
        return_annotation:
          type: string
        comment:
          type: string
    ConfigurationResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
          properties:
            configuration:
              type: object
              additionalProperties: true
    DirectoryResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
          properties:
            file_directory:
              type: object
              additionalProperties:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: object
    ProjectListResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
          properties:
            project_list:
              type: array
              items:
                type: string
    ConfigurationRequest:
      type: object
      properties:
        content:
          type: object
          additionalProperties: true
    FileModificationRequest:
      type: object
      properties:
        content:
          type: string
          description: 文件内容
    FileContentResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
          properties:
            content:
              type: string
    MissionStatusResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: object
          additionalProperties:
            type: boolean
    MissionRequest:
      type: object
      additionalProperties:
        type: object
        additionalProperties:
          type: object
          properties:
            STRATEGY_QUEUE:
              type: array
              items:
                type: object
                required: [ID, FUNC, ARGS]
                properties:
                  ID:
                    type: string
                  FUNC:
                    type: string
                  ARGS:
                    type: object
                    additionalProperties: 
                      type: string
            ITER:
              type: boolean
            GET_OUTPUT:
              type: array
              items:
                type: string
    MissionResponse:
      type: object
      properties:
        message:
          type: array
          items:
            type: string
        status:
          type: object
          additionalProperties:
            type: boolean
        data:
          type: object
          additionalProperties:
            type: string
    SystemStatusResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
          properties:
            CPU:
              type: object
              properties:
                load:
                  type: string
                temperature:
                  type: string
                power:
                  type: string
            RAM:
              type: object
              properties:
                load:
                  type: string
            GPU:
              type: object
              properties:
                load:
                  type: string
                temperature:
                  type: string
                power:
                  type: string
            HDD:
              type: object
              properties:
                load:
                  type: string

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
    
    SuccessResponse:
      type: object
      properties:
        message:
          type: string